+++ src/main/kotlin/com/github/warningimhack3r/npmupdatedependencies/NUDConstants.kt
@@ -4,4 +4,5 @@ object NUDConstants {
     const val NPMJS_REGISTRY = "https://registry.npmjs.com"
     const val PACKAGE_JSON = "package.json"
     val dependenciesKeys = listOf("dependencies", "devDependencies")
+    const val PACKAGE_MANAGER_KEY = "packageManager"
 }